+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- fix: Scope values should not override Event values (#446)
 - feat: Extend User inteface by adding Data, Name and Segment (#483)
 
 ## 0.14.0